pyproject.toml
[tool.poetry]
name = "general-pesquisa"
version = "0.1.0"
description = ""
authors = ["Luciano Borba <luhborbafilho@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.32.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
